+++ cursor/src/main/kotlin/cursors/Cursor.kt
@@ -7,13 +7,10 @@ import cursors.calendar.feature_range
 import cursors.context.*
 import cursors.io.*
  import vec.macros.*
-import vec.util.logDebug
 import java.time.LocalDate
 import java.util.*
 import kotlin.Comparator
 import kotlin.coroutines.CoroutineContext
-import kotlin.math.max
-import kotlin.math.sqrt
 
 /**
  * cursor is approximately the pandas Dataframe with some shortcuts
@@ -183,78 +180,6 @@ fun Cursor.pivot(
     }
 }
 
-
-inline fun Cursor.group(
-    /**these columns will be preserved as the cluster key.
-     * the remaining indexes will be aggregate
-     */
-    vararg axis: Int
-): Cursor = let { orig ->
-    val clusters =  groupClusters(axis)
-    val masterScalars = orig.scalars
-    Cursor(clusters.size) { cy ->
-        val cluster = clusters[cy]
-        val cfirst = cluster.first()
-        RowVec(masterScalars.first) { ix: Int ->
-            when (ix) {
-                in axis -> (orig at (cfirst))[ix]
-                else -> Vect0r(cluster.size) { iy: Int -> (orig at (cluster[iy]))[ix].first } t2 masterScalars[ix].`⟲`
-            }
-        }
-    }
-}
-
-inline fun Cursor.group(
-    /**these columns will be preserved as the cluster key.
-     * the remaining indexes will be aggregate
-     */
-    axis: IntArray,
-    crossinline reducer: ((Any?, Any?) -> Any?)
-): Cursor = run {
-    val clusters =  groupClusters(axis)
-    val masterScalars = scalars
-    val xSize = masterScalars.first
-    Cursor(clusters.size) { cy ->
-        val acc1 = arrayOfNulls<Any?>(xSize)
-        val cluster = clusters[cy]
-        val keyIndex = cluster.first()
-        val valueIndices = acc1.indices - axis.toTypedArray()
-
-        for (i in cluster) {
-            val value = (this at (i)).left
-            for (valueIndex in valueIndices)
-                acc1[valueIndex] = reducer(acc1[valueIndex], value[valueIndex])
-        }
-        RowVec(masterScalars.first) { ix: Int ->
-            if (ix in axis) {
-                (this at (keyIndex))[ix]
-            } else acc1[ix] t2 masterScalars[ix].`⟲`
-        }
-    }
-}
-
-inline fun Cursor.groupClusters(
-    axis: IntArray,
-    clusters: MutableMap<List<Any?>, MutableList<Int>> = linkedMapOf()
-) = run {
-    System.err.println("--- groupClusters")
-    keyClusters(axis, clusters)
-    clusters.values α MutableList<Int>::toIntArray
-}
-
-inline fun Cursor.keyClusters(
-    axis: IntArray,
-    clusters: MutableMap<List<Any?>, MutableList<Int>>
-): MutableMap<List<Any?>, MutableList<Int>> = clusters.apply {
-    val cap = max(8, sqrt(size.toDouble()).toInt())
-
-    forEachIndexed { iy: Int, row: RowVec ->
-        row[axis].left.toList().let {
-            getOrPut(it) { ArrayList(cap) } += iy
-        }
-    }
-    logDebug { "cap: $cap keys:${clusters.size to clusters.keys}" }
-}
 
 /**
  * this is a helper for comparing keys.
