+++ cursor/src/test/resources/List_of_largest_cities.html
@@ -1,0 +1,1 @@
+../../../../vector-like/src/test/resources/List_of_largest_cities.html
