+++ vector-like/src/main/kotlin/vec/macros/Twop13.kt
@@ -45,21 +45,21 @@ interface Pai2<F, S> {
     @Suppress("OVERRIDE_BY_INLINE")
     companion object {
         inline operator fun <reified F, reified S> invoke(first: F, second: S): Pai2<F, S> =
-            object : Pai2<F, S> {
-                override inline val first get() = first
-                override inline val second get() = second
-            }
+                object : Pai2<F, S> {
+                    override inline val first get() = first
+                    override inline val second get() = second
+                }
 
         /**
          * Pair copy ctor conversion
          */
         inline operator fun <reified F, reified S, reified P : Pair<F, S>, reified R : Pai2<F, S>> invoke(p: P): R =
-            p.run { (first t2 second) as R }
+                p.run { (first t2 second) as R }
     }
 }
-
+/*
 inline operator fun <reified F, reified S> Pai2<F, S>.component2(): S = second
-inline operator fun <reified F, reified S> Pai2<F, S>.component1(): F = first
+inline operator fun <reified F, reified S> Pai2<F, S>.component1(): F = first*/
 
 /**inheritable version of triple
  */
@@ -75,14 +75,14 @@ interface Tripl3<out F, out S, out T>/* : Pai2<F, S> */ {
 
     companion object {
         inline operator fun <reified F, reified S, reified T> invoke(first: F, second: S, third: T): Tripl3<F, S, T> =
-            object : Tripl3<F, S, T>/*, Pai2<F, S> by Pai2(f, s)*/ {
-                override inline val first get() = first
-                override inline val second get() = second
-                override inline val third get() = third
-            }
+                object : Tripl3<F, S, T>/*, Pai2<F, S> by Pai2(f, s)*/ {
+                    override inline val first get() = first
+                    override inline val second get() = second
+                    override inline val third get() = third
+                }
 
         inline operator fun <reified F, reified S, reified T> invoke(p: Triple<F, S, T>) =
-            p.let { (f, s, t) -> Tripl3(f, s, t) }
+                p.let { (f, s, t) -> Tripl3(f, s, t) }
     }
 
 }
@@ -128,6 +128,7 @@ inline class Tw1nt(val ia: IntArray) : Tw1n<Int> {
     override inline val second get() = ia[1]
 
 }
+
 /**
  * long-type specific twin with primitive array backing store.
  */
@@ -146,29 +147,29 @@ inline fun <reified T : Long> Tw1n(first: T, second: T) = Twln(longArrayOf(first
 
 
 inline infix fun <reified X, reified Y, Z, P : Pai2<X, Y>, U : Hand1l<X>, T : Hand1l<Y>> U.asLeft(u: T) =
-    (first t2 u.first) as P
+        (first t2 u.first) as P
 
 inline infix fun <reified X, reified Y, reified Z, P : Pai2<Y, Z>, U : Hand1l<X>, T : Tripl3<X, Y, Z>> U.asLeft(
-    p: P
+        p: P,
 ) = Tripl3(first, p.first, p.second)
 
 inline infix fun <reified X, reified Y, reified Z, U : Hand1l<X>, T : Tripl3<X, Y, Z>> U.asLeft(t: T) =
-    Qu4d(first, t.first, t.second, t.third)
+        Qu4d(first, t.first, t.second, t.third)
 
 inline infix fun <reified F, reified S> F.t2(s: S) = Pai2(this, s)
 inline infix fun <reified F, reified S, reified T> Pai2<F, S>.t3(t: T) = let { (f: F, s) ->
     Tripl3(
-        f,
-        s,
-        t
+            f,
+            s,
+            t
     )
 }
 
 inline infix fun <reified F, reified S, reified T, P : Pair<F, S>> P.t3(t: T) =
-    let { (a, b) -> Tripl3(a, b, t) }
+        let { (a, b) -> Tripl3(a, b, t) }
 
 inline infix fun <reified A, reified B, reified C, reified D> Tripl3<A, B, C>.t4(d: D) =
-    let { (a: A, b: B, c: C) -> Qu4d(a, b, c, d) }
+        let { (a: A, b: B, c: C) -> Qu4d(a, b, c, d) }
 
 
 /**inheritable version of quad that also provides its first three as a triple. */
@@ -185,30 +186,27 @@ interface Qu4d<F, S, T, Z> {
      */
     val quad: Quad<F, S, T, Z>
         get() = Quad(
-            first
-            , second
-            , third
-            , fourth
+                first, second, third, fourth
         )
 
     companion object {
         inline operator fun <reified F, reified S, reified T, reified Z> invoke(
-            first: F,
-            second: S,
-            third: T,
-            fourth: Z
+                first: F,
+                second: S,
+                third: T,
+                fourth: Z,
         ): Qu4d<F, S, T, Z> =
-            object : Qu4d<F, S, T, Z> {
-                override inline val first get() = first
-                override inline val second get() = second
-                override inline val third get() = third
-                override inline val fourth get() = fourth
-            }
+                object : Qu4d<F, S, T, Z> {
+                    override inline val first get() = first
+                    override inline val second get() = second
+                    override inline val third get() = third
+                    override inline val fourth get() = fourth
+                }
 
         inline operator fun <reified F, reified S, reified T, reified Z> invoke(p: Quad<F, S, T, Z>) =
-            p.let { (f, s, t, z) ->
-                Qu4d(f, s, t, z)
-            }
+                p.let { (f, s, t, z) ->
+                    Qu4d(f, s, t, z)
+                }
 
         inline operator fun <reified F, reified S, reified T, reified Z> invoke(p: Array<*>) = p.let { (f, s, t, z) ->
             @Suppress("UNCHECKED_CAST")
